[REFACTOR] - 내역 리사이클러뷰 뷰타입 상수로 관리하도록 수정
