+++ app/src/main/java/com/nimok97/accountbook/common/HistoryViewType.kt
@@ -1,0 +1,5 @@
+package com.nimok97.accountbook.common
+
+const val HISTORY_HEADER = "HEADER"
+const val HISTORY_CONTENT = "CONTENT"
+const val HISTORY_FOOTER = "FOOTER"
